+++ tests/common.rs
@@ -63,7 +63,7 @@ where
     println!("\texhaustive search:");
     let now = std::time::Instant::now();
     let exp_result = exhaustive_search(&mut exp_selection, &mut metric);
-    let exp_change = metric.drain(&exp_selection);
+    let exp_change = metric.drain(&exp_selection.compute_view());
     let exp_result_str = result_string(&exp_result.ok_or("no possible solution"), exp_change);
     println!(
         "\t\telapsed={:8}s result={}",
@@ -73,7 +73,7 @@ where
     // bonus check: ensure replacement fee is respected
     if exp_result.is_some() {
         let selected_value = exp_selection.selected_value();
-        let drain = metric.drain(&exp_selection);
+        let drain = metric.drain(&exp_selection.compute_view());
         let target_value = target.value();
         let replace_fee = params
             .replace
@@ -89,7 +89,7 @@ where
     let now = std::time::Instant::now();
     let mut bnb_metric = metric.clone();
     let result = bnb_search(&mut selection, metric, usize::MAX);
-    let change = bnb_metric.drain(&selection);
+    let change = bnb_metric.drain(&selection.compute_view());
     let result_str = result_string(&result, change);
     println!(
         "\t\telapsed={:8}s result={}",
@@ -113,7 +113,7 @@ where
 
             // bonus check: ensure replacement fee is respected
             let selected_value = selection.selected_value();
-            let drain = bnb_metric.drain(&selection);
+            let drain = bnb_metric.drain(&selection.compute_view());
             let target_value = target.value();
             let replace_fee = params
                 .replace
@@ -159,12 +159,12 @@ where
     print_candidates(&params, &init_cs);
 
     for (cs, _) in ExhaustiveIter::new(&init_cs).into_iter().flatten() {
-        if let Some(lb_score) = metric.bound(&cs) {
+        if let Some(lb_score) = metric.bound(&cs.compute_view()) {
             // This is the branch's lower bound. In other words, this is the BEST selection
             // possible (can overshoot) traversing down this branch. Let's check that!
 
-            if let Some(score) = metric.score(&cs) {
-                let has_change = metric.drain(&cs).is_some();
+            if let Some(score) = metric.score(&cs.compute_view()) {
+                let has_change = metric.drain(&cs.compute_view()).is_some();
                 prop_assert!(
                     score >= lb_score,
                     "checking branch: selection={} score={} change={} lb={}",
@@ -180,9 +180,10 @@ where
                 .flatten()
                 .filter(|(_, inc)| *inc)
             {
-                if let Some(descendant_score) = metric.score(&descendant_cs) {
-                    let parent_has_change = metric.drain(&cs).is_some();
-                    let descendant_has_change = metric.drain(&descendant_cs).is_some();
+                if let Some(descendant_score) = metric.score(&descendant_cs.compute_view()) {
+                    let parent_has_change = metric.drain(&cs.compute_view()).is_some();
+                    let descendant_has_change =
+                        metric.drain(&descendant_cs.compute_view()).is_some();
                     prop_assert!(
                         descendant_score >= lb_score,
                         "
@@ -364,7 +365,7 @@ where
         .enumerate()
         .inspect(|(i, _)| rounds = *i)
         .filter(|(_, (_, inclusion))| *inclusion)
-        .filter_map(|(_, (cs, _))| metric.score(&cs).map(|score| (cs, score)));
+        .filter_map(|(_, (cs, _))| metric.score(&cs.compute_view()).map(|score| (cs, score)));
 
     for (child_cs, score) in iter {
         match &mut best {
@@ -504,11 +505,11 @@ pub fn compare_against_benchmarks<M: BnbMetric + Clone>(
             let cmp_benchmarks = cmp_benchmarks
                 .into_iter()
                 .filter_map(|cs| {
-                    let score = metric.clone().score(&cs)?;
+                    let score = metric.clone().score(&cs.compute_view())?;
                     Some((cs, score))
                 })
                 .collect::<Vec<_>>();
-            let sol_score = metric.score(&sol);
+            let sol_score = metric.score(&sol.compute_view());
 
             for (_bench_id, (mut bench, bench_score)) in cmp_benchmarks.into_iter().enumerate() {
                 prop_assert!(
@@ -550,7 +551,7 @@ fn randomly_satisfy_target<'a, R: rand::Rng>(
     while let Some(next) = cs.unselected_indices().choose(rng) {
         cs.select(next);
         if cs.is_funded() {
-            let curr_score = metric.score(&cs);
+            let curr_score = metric.score(&cs.compute_view());
             if let Some(last_score) = last_score {
                 if curr_score.is_none() || curr_score.unwrap() > last_score {
                     break;